backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Werkzeug==3.0.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
PyJWT==2.8.0
pymysql==1.1.0
flask-cors==4.0.0
click==8.1.7
markupsafe==2.1.3
Jinja2==3.1.2
itsdangerous==2.1.2
blinker==1.7.0